+++ MEIClient/metee.cmake
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
-# Copyright (C) 2019-2021 Intel Corporation
-cmake_minimum_required (VERSION 3.1)
+# Copyright (C) 2019-2024 Intel Corporation
+cmake_minimum_required (VERSION 3.15)
 
 find_package("metee")
 if(NOT METEE_FOUND)