+++ UNS/GMS_COMMON/Tools.cpp
@@ -1,9 +1,10 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /*
- * Copyright (C) 2010-2023 Intel Corporation
+ * Copyright (C) 2010-2024 Intel Corporation
  */
 #include "Tools.h"
 #include <time.h>
+#include <vector>
 #include <string>
 #include <iomanip>
 #include <iostream>
@@ -12,10 +13,13 @@
 #include <cctype>
 #ifdef WIN32
 #include <Userenv.h>
+#include <Wbemidl.h>
+#include <atlbase.h>
 #else
 #include <netdb.h>
 #include <unistd.h>
 #include <string.h>
+#include <sys/stat.h>
 #endif // WIN32
 
 std::string getDateTime()
@@ -199,3 +203,99 @@ recode:
 	return true;
 }
 #endif // WIN32
+
+#ifdef WIN32
+bool MEIEnabled(std::wstring &err)
+{
+	bool meiEnabled = false;
+	IWbemLocator* loc = NULL;
+	IWbemServices* svc = NULL;
+	IWbemClassObject* obj = NULL;
+	ULONG uReturn = 0;
+	std::wstringstream ss;
+
+	try
+	{
+		HRESULT hres = CoCreateInstance(__uuidof(WbemLocator), 0, CLSCTX_INPROC_SERVER, __uuidof(IWbemLocator), (LPVOID*)&loc);
+
+		if (!FAILED(hres))
+		{
+			// Connect to the root\cimv2 namespace with
+			// the current user and obtain pointer pSvc
+			// to make IWbemServices calls.
+			hres = loc->ConnectServer(CComBSTR(L"ROOT\\CIMV2"), NULL, NULL, 0, NULL, 0, 0, &svc);
+			if (!FAILED(hres))
+			{
+				IEnumWbemClassObject* enumerator = NULL;
+				hres = svc->ExecQuery(CComBSTR(L"WQL"),
+					CComBSTR(L"SELECT Status FROM Win32_PnPEntity where Caption = \'Intel(R) Management Engine Interface \' or Caption = \'Intel(R) Management Engine Interface #1\'"),
+					WBEM_FLAG_FORWARD_ONLY | WBEM_FLAG_RETURN_IMMEDIATELY, NULL, &enumerator);
+				if (!FAILED(hres))
+				{
+					if (enumerator)//we have only one instance
+					{
+						hres = enumerator->Next(WBEM_INFINITE, 1, &obj, &uReturn);
+						if (uReturn != 0)
+						{
+							VARIANT vtProp;
+							hres = obj->Get(L"Status", 0, &vtProp, 0, 0);
+							if (SUCCEEDED(hres))
+							{
+								if (wcscmp(vtProp.bstrVal, L"OK") == 0)
+								{
+									meiEnabled = true;
+								}
+							}
+							else
+							{
+								ss << L"isMEIEnabled() failed to get status " << hres;
+							}
+						}
+						else
+						{
+							ss << L"isMEIEnabled() failed to enumerate device " <<  hres;
+						}
+					}
+				}
+				else
+				{
+					ss << L"isMEIEnabled() failed to connect to exec WMI query";
+				}
+			}
+			else
+			{
+				ss << L"isMEIEnabled() failed to connect to WMI server";
+			}
+		}
+		else
+		{
+			ss << L"isMEIEnabled() failed in CoCreateInstance()";
+		}
+	}
+	catch (const ATL::CAtlException& e)
+	{
+		ss << "isMEIEnabled() AtlException hr = 0x" << std::hex << e.m_hr;
+	}
+	if (svc != NULL) svc->Release();
+	if (loc != NULL) loc->Release();
+	if (obj != NULL) obj->Release();
+	err = ss.str();
+	return meiEnabled;
+}
+#else
+bool MEIEnabled(std::wstring& err)
+{
+	struct stat buf;
+	static const std::vector<std::string> devnode =
+	{ "/dev/mei", "/dev/mei0", "/dev/mei1", "/dev/mei2", "/dev/mei3" };
+
+	for (std::vector<std::string>::const_iterator it = devnode.begin();
+		it != devnode.end(); ++it)
+	{
+		if (!lstat(it->c_str(), &buf))
+			return true;
+	}
+	err = L"Mei device is not found";
+	return false;
+}
+#endif //WIN32