+++ conanfile.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 # SPDX-License-Identifier: Apache-2.0
-# Copyright (C) 2021-2023 Intel Corporation
+# Copyright (C) 2021-2024 Intel Corporation
 from conans import ConanFile
 import os
 
@@ -13,10 +13,10 @@ class LMSConan(ConanFile):
 
     def requirements(self):
         if self.settings.os == "Windows":
-            self.requires("libxml2/2.12.2@mesw/stable")
-            self.requires("ACE/7.1.2@mesw/stable")
+            self.requires("libxml2/2.12.7@mesw/stable")
+            self.requires("ACE/7.1.3@mesw/stable")
         else:
             if os.environ.get("LOCAL_ACE", None):
-                self.requires("ACE/7.1.2@mesw/stable")
+                self.requires("ACE/7.1.3@mesw/stable")
         if self.options.build_tests:
             self.requires("gtest/1.10.0@mesw/stable")